fix(datepicker) month view display bug
The month was not rendering the correct starting day in certain cases.  For example: on the Angular Strap datepicker demo: February 2015 is rendered as starting on a Saturday when it actually starts on a Sunday.

fix(datepicker) update dateFormat test

Previous datepicker bug fix caused January 26th to become visible in this month so the day the test selected has changed.

fix(datepicker) update options-startWeek-bis test