fix(typeahead): show dropdown on focus when minLegth is 0
The first time the input control receives focus and minLegth was 0, the options dropdown was not showing. If you typed a character and then deleted it, then the dropdown would always show on focus. Turns out that the first time, the view value was undefined, so the internal $isVisible value was returning false. Fixed it by defaulting undefined display values to empty string.

Might be the cause for #1164.