test(datepicker): add test to check days order in datepicker
Added a new gulp task called 'test:timezone' that sets the environment timezone for PhantomJS to use. It expects a parameter on the command line with the name of the timezone to use, like: 'gulp test:timezone --America/Sao_Paulo'.
Using this new task, the new test that checks for days order in datepicker fails for America/Sao_Paulo because of changes in Daylight Saving Time - from what I could understand, Brazil changes the time at midnight and the datepicker uses dates with time set to midnight, so when it increments one day, it goes back to the previous day, which caused repeated days to show up on the calendar.