feat(typeahead): add $id property to $typeahead service instance
Add an $id property to each $typeahead instance to allow for event handlers to identify which element is triggering the event.
The value can be set using the element's id attribute or by passing the desired value in the config object when using the $typeahead service directly.
Adds a third parameter to the '.select' event with the value of $typeahead instance that triggered the event (keeps existing parameters order so it doesn't break compatibility)